requirements.txt
google-genai
googlemaps
python-dotenv
requests
